requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
gunicorn==21.2.0

openai==1.50.0
httpx==0.25.2

pdfplumber==0.9.0

google-cloud-vision==3.4.4
google-cloud-storage==2.10.0
google-auth==2.23.4

Flask-Session==0.5.0
redis==5.0.0

qdrant-client==1.9.1
python-dotenv==1.0.1
